The Full LSTM method with independent RNNs.
- ExRecCNOTFullLSTM is rrevised not to flag reuse=True.
This improves the results and is the way to go.
- Results are better for larger training data sets. So with higher
iteration we should be able to get much better results.
- ExRecCNOTFullLSTMS2S attempts at comparing the full error sequence
to the error sequence given the two outputs of the 2 replications of
the LSTM cell. This is not working at the moment due to memory
issues. I don't expect this to be successful.